+++ presta/app/__init__.py
@@ -3,7 +3,10 @@ import os
 from celery import Celery
 
 app = Celery('app',
-             include=['app.tasks'])
+             include=['app.tasks',
+                      'app.cron',
+                      'app.events',
+                      ])
 
 # Set default configuration module name
 os.environ.setdefault('CELERY_CONFIG_MODULE', 'app.celeryconfig')
